+++ apps/builder/app/shared/canvas-api.ts
@@ -4,11 +4,11 @@ import { monitorForExternal } from "@atlaskit/pragmatic-drag-and-drop/external/a
 import { createRecursiveProxy } from "@trpc/server/shared";
 import invariant from "tiny-invariant";
 import { $canvasIframeState } from "./nano-states";
-import { getElementAndAncestorInstanceTags } from "~/canvas/instance-selected";
 import { detectSupportedFontWeights } from "~/canvas/shared/font-weight-support";
 import {
   getBrowserStyle,
   calculateUnitSizes,
+  getElementAndAncestorInstanceTags,
 } from "~/canvas/features/webstudio-component/get-browser-style";
 
 const apiWindowNamespace = "__webstudio__$__canvasApi";
